Major rewrite
Updated the app to use a lot of the code also used for yabs. This should
make it easier to update and modify the site. There is also basic
support for internationalization.

The design is now also easier to modify and it uses Bootstrap, select2
for tag suggestions, a new version of plupload for file uploading and
bluimp Gallery for the lightbox viewer.

Login system has been changed to use Google Sign In instead of OpenID
since OpenID is no longer supported by Google.

The database layout has changed slightly. Older versions of the site can
be updated using https://gist.github.com/Trellmor/45d21997ccaf307e44f2